PXF Docker environment for tests running (#33)
* PXF Docker environment for tests running

The goal of this patch is to begin automation of PXF testing. It contains
scripts and images for Docker environment, like it's done for other ADB
components and extensions. Existing upstream concourse scripts were reused where
it was possible. For now, we don't focus on hadoop integration testing, but do
unit testing, fdw and external-table regression testing.

* ADBDEV-3356: Fix ClassCastException issue in DemoTextResolver

If the output format of the external table is 'TEXT' we return DataInputStream instead of byte array

---------

Co-authored-by: Roman Zolotov <rzolotov@arenadata.io>