ADBDEV-4011: Fix DemoTextResolver
In the previous releases, if an external table has 'TEXT' output format a text row data was passed as a single field. In this release, the logic was changed. The new annotation was introduced. This annotation is used for marking Plugins as capable of handling InputStream with raw data. DemoTextResolver can handle InputStream, but it didn't have this annotation.