ADBDEV-5283: Bump Springboot to 2.7.18
We used the Springboot version 2.4.3 because of the bug with queries which contain LIMIT statement. The bug is described here https://github.com/greenplum-db/pxf/issues/947 and comes (as figured out) from spring-mvc.

In Springboot version 2.7.18 the bug was resolved. Details are here: https://github.com/greenplum-db/pxf/pull/1105

Now, we can bump the version 2.4.3 to 2.7.18.