TLS options implemented for PXF FDW (#146)
TLS options implemented for PXF FDW

The requirement for using TLS is implemented for PXF FDW. The
communication takes place between the segment node and the PXF server and is
pointless without the PXF server. The testing shall be done in conjunction with
PXF server changes.

To test TLS support there is a new test container Dockerfile_ssl, which adds
PXF_SSL variables and additional SSL-related setup.

The requirement for using TLS is implemented for PXF external tables.
Communication takes place between the segment node and the PXF server and is
pointless without the PXF server. Testing should be done in conjunction with PXF
server changes and is out of scope for this commit.

This commit also has merged related commit #141:

TLS options implemented for PXF external tables

The requirement for using TLS is implemented for PXF external tables.
Communication takes place between the segment node and the PXF server and is
pointless without the PXF server. Testing should be done in conjunction with PXF
server changes and is out of scope for this commit.

PXF external tables are controlled by environment variables. The following
options are affected by this commit:

    PXF_PROTOCOL: Type string, default http. Defines the protocol for
    communication with the PXF server. Possible values are http and https.
    PXF_SSL_CACERT: Type string, default
    /home/gpadmin/arenadata_configs/cacert.pem. Specifies the absolute path to the
    CA certificate that signed the server certificate. Refers to CURLOPT_CAINFO.
    PXF_SSL_CERT: Type string, default client.pem. Specifies the path to the
    client certificate. Refers to CURLOPT_SSLCERT.
    PXF_SSL_CERT_TYPE: Type string, default pem. Defines the type of the client
    certificate. Refers to CURLOPT_SSLCERTTYPE.
    PXF_SSL_KEY: Type string, default empty. Defines the path to the client's SSL
    private key. Refers to CURLOPT_SSLKEY.
    PXF_SSL_KEYPASSWD: Type string, default empty. Defines the password for the
    client SSL private key. Refers to CURLOPT_KEYPASSWD.
    PXF_SSL_VERIFY_PEER: Type long, default 1. Defines whether the client should
    verify the server certificate. Refers to CURLOPT_SSL_VERIFYPEER.