Allow interrupts in finish_upload (#173)
In some circumstances, such as the unavailability of a external systems, we may
be stuck for a long time in the finish_upload function. This happened because we
are waiting for final PXF response and connections to be closed. For this, we
call the curl_multi_perform function in a loop. Since there is nothing to
receive or send, it immediately exits, while informing us that we still have
open connections. After that, we will stay in this loop until PXF finally
reports something. The loop is busy spinning and burns CPU. It is also not
interruptible: Greenplum cancel requests weren’t processed promptly, so users
couldn’t cancel stuck backends.

The proposed solution lies in including CHECK_FOR_INTERRUPTS() inside the loop
so cancels are honored promptly, and in using curl_multi_wait() between
curl_multi_perform() calls to wait for socket activity or a short timeout, 
liminating the busy-spin and letting libcurl make progress when data arrives.

Co-authored-by: Alexander Kondakov <a.kondakov@arenadata.io>